Finland/requirements.txt
curl_cffi>=0.7.0
python-dotenv>=1.0.0
lxml>=5.0.0
requests>=2.31.0
openpyxl>=3.1.0
openai>=1.68.0
markdownify
beautifulsoup4>=4.12.0
